data-collection/requirements.txt
# Data Collection Scripts
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
